[fixed] harvester radials
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@2397 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/packets/object/RadialMenuItem.h
@@ -121,7 +121,7 @@ public:
 		return returnItem;
 	}
 
-	RadialMenuItem* getItemByRadialID(int index) {
+	RadialMenuItem* getItemByRadialID(uint8 index) {
 		RadialMenuItem* returnItem = NULL;
 
 		for (int i = 0; i < children.size(); ++i) {
